pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ark_roulette"
version = "1.2.0"
description = "一个用于 NoneBot 的插件，提供明日方舟干员筛选与随机选择功能，适用于群聊娱乐。"
authors = ["星之游"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/xingzhiyou/nonebot-plugin-ark-roulette"
repository = "https://github.com/xingzhiyou/nonebot-plugin-ark-roulette"
keywords = ["nonebot", "plugin", "arknights", "roulette", "chatbot"]

packages = [{ include = "nonebot_plugin_ark_roulette", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
nonebot2 = "^2.0.2"
nonebot-adapter-onebot = "^2.0.0"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^23.9.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"